Moved main to separate file and added notes
